+++ ds2m/caio_palermo_lemos/css/chats/chatsContainerStyle.css
@@ -22,8 +22,8 @@ span:hover {
 
 .chat {
   display: grid;
-  grid-template-areas: "image content";
-  grid-template-columns: auto 1fr;
+  grid-template-areas: "image content time";
+  grid-template-columns: auto 1fr auto;
   align-items: center;
   margin-right: 12px;
 }
