+++ ds2m/caio_palermo_lemos/js/search-contact.js
@@ -1,32 +1,26 @@
-"use strict";
+// "use strict";
 
-//BARRA DE PESQUISAR
-// const buscar = function () {
-//   const filterInput = document.querySelector(".filter__input");
-//   const chats = document.querySelectorAll(".chat");
+const filter = document.querySelector(".filter__input");
+const targets = document.querySelectorAll(".chat");
 
-//   const handleKey = function (event) {
-//     console.log(event.key);
-//     filterInput.value = "";
-//     if (filterInput.value != "") {
-//       contatos.forEach((contato) => {
-//         filterInput.value = "Mark";
-//         if (contato.name.startsWith(filterInput.value)) {
-//           chats.forEach((item) => {
-//             let nome = item.innerText.split(`\n`)[0];
-//             item.classList.add("none");
+const searchContact = function () {
+  const handleKey = function (event) {
+    const value = event.target.value;
+    if (value.trim().length) {
+      const regex = new RegExp(`^(${value})`, "gi");
+      targets.forEach((target) => {
+        const targetValue = target.innerText.split(`\n`)[0];
+        if (regex.test(targetValue)) {
+        } else {
+          target.classList.add("none");
+        }
+      });
+    } else {
+      targets.forEach((target) => target.classList.remove("none"));
+    }
+  };
 
-//             if (nome == contato.name) {
-//               console.log("O nome é: " + nome);
-//               item.classList.remove("none");
-//             }
-//           });
-//         }
-//       });
-//     }
-//   };
+  filter.addEventListener("keyup", handleKey);
+};
 
-//   filterInput.addEventListener("keypress", handleKey);
-// };
-
-// buscar();
+searchContact();
